+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Release date: ?
  - added multiple set type methods in Collection\Structure\DataType
  - setProps() of Backpack\AppBuilder now accept an array or an instance of Blueprint\Collection\Dictionary
  - added Dictionary Blueprint to class Collection\DotNotationCollection
+ - added method stackIfTrue() for conditional stacking in Bedrock\Application\Application 
 
 VERSION 4.0-BETA1
 -----------